Renamed package to follow java conventions
