+++ application/pom.xml
@@ -3,7 +3,7 @@
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <modelVersion>4.0.0</modelVersion>
 
-    <groupId>com.github.pieter-groenendijk</groupId>
+    <groupId>com.github.pietergroenendijk</groupId>
     <artifactId>library-management</artifactId>
     <version>1.0-SNAPSHOT</version>
