+++ application/src/main/java/com/github/pietergroenendijk/App.java
@@ -1,4 +1,4 @@
-package com.mycompany.app;
+package com.github.pietergroenendijk;
 
 /**
  * Hello world!
