+++ application/src/test/java/com/github/pietergroenendijk/AppTest.java
@@ -1,0 +1,1 @@
+package com.github.pietergroenendijk;
